+++ backend/Models/SavingGoal.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Details { get; set; } = String.Empty;
-        public int Goal { get; set; }
+        public double Goal { get; set; }
         public int? FixedContribution { get; set; }
         public int? Periodicity { get; set; }
         public List<GoalContribution> GoalContributions { get; set; } = new List<GoalContribution>();
